BACKEND-001: Fix backend compilation errors and add missing repositories
## Changes Made:

### New Repositories:
- Created VideoRepository with search and filtering methods
- Created PlaylistRepository with search and filtering methods

### Model Enhancements:
- Channel: Added name, description, thumbnailUrl, subscribers, videoCount fields
- Playlist: Added title, description, thumbnailUrl, itemCount fields
- Video: Added helper method getCategory()
- Category: Added slug and parentId fields with getters

### Repository Enhancements:
- ChannelRepository: Added findByCategoryOrderBySubscribersDesc(), findAllByOrderBySubscribersDesc(), searchByName()
- All repositories implement proper Firestore query patterns

### Service Layer Fixes:
- PublicContentService: Added CategoryDto import
- Fixed type mismatch for subscribers (Long vs Integer)
- Fixed Timestamp to LocalDateTime conversion for Video.uploadedAt
- Added ExecutionException and InterruptedException handling
- Simplified Video.toDto() to avoid checked exceptions in streams

### Controller Updates:
- PublicContentController: Added throws clauses for all endpoints

### Build Status:
✅ Backend builds successfully: ./gradlew clean build -x test

### Blocker Identified:
Backend requires Firebase configuration to run:
- Firebase service account JSON needed
- Firestore database must be enabled
- See BACKEND_STATUS.md and backend/FIREBASE_SETUP.md for details

## Next Steps:
1. User decision: Set up Firebase OR continue with mock data
2. See BACKEND_STATUS.md for options

+++ backend/src/main/java/com/albunyaan/tube/dto/ContentItemDto.java
@@ -24,7 +24,7 @@ public class ContentItemDto {
     private Long viewCount;
 
     // Channel-specific
-    private Integer subscribers;
+    private Long subscribers;
     private Integer videoCount;
 
     // Playlist-specific
@@ -35,7 +35,7 @@ public class ContentItemDto {
 
     // Builder pattern
     public static ContentItemDto channel(String id, String name, String category,
-                                         Integer subscribers, String description,
+                                         Long subscribers, String description,
                                          String thumbnailUrl, Integer videoCount) {
         ContentItemDto dto = new ContentItemDto();
         dto.id = id;
@@ -160,11 +160,11 @@ public class ContentItemDto {
         this.viewCount = viewCount;
     }
 
-    public Integer getSubscribers() {
+    public Long getSubscribers() {
         return subscribers;
     }
 
-    public void setSubscribers(Integer subscribers) {
+    public void setSubscribers(Long subscribers) {
         this.subscribers = subscribers;
     }
